# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Entitlements API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/entitlements/viewers": {
            "post": {
                "summary": "Register new viewer",
                "operationId": "Modules\\Entitlements\\Http\\Controllers\\ViewersController::registerViewer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Id of viewer which should be registered",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/entitlements/viewers/{viewerId}/plans": {
            "put": {
                "summary": "Register that viewer purchased a plan",
                "operationId": "Modules\\Entitlements\\Http\\Controllers\\ViewersController::registerViewerPurchasedPlan",
                "parameters": [
                    {
                        "name": "viewerId",
                        "in": "path",
                        "description": "Id of viewer who purchased a plan",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "planId",
                                    "expiresAt"
                                ],
                                "properties": {
                                    "planId": {
                                        "description": "Id of purchased plan",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                                    },
                                    "expiresAt": {
                                        "description": "Moment of plan expiration",
                                        "type": "string",
                                        "format": "datetime",
                                        "example": "2021-01-01 00:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/entitlements/viewers/{viewerId}/videos": {
            "put": {
                "summary": "Register that viewer purchased a pay per view video",
                "operationId": "Modules\\Entitlements\\Http\\Controllers\\ViewersController::registerViewerPurchasedPayPerViewVideo",
                "parameters": [
                    {
                        "name": "viewerId",
                        "in": "path",
                        "description": "Id of viewer who purchased a video",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "videoId"
                                ],
                                "properties": {
                                    "videoId": {
                                        "description": "Id of purchased video",
                                        "type": "string",
                                        "format": "uuid",
                                        "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/resources/broadcasters/{broadcasterId}/plans": {
            "get": {
                "summary": "Get broadcasters plans",
                "operationId": "Modules\\Resources\\Http\\Controllers\\PlansController::getBroadcastersPlans",
                "parameters": [
                    {
                        "name": "broadcasterId",
                        "in": "path",
                        "description": "Id of broadcaster",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Broadcastesrs plans ids",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "plansIds": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "format": "uuid"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Add new plan",
                "operationId": "Modules\\Resources\\Http\\Controllers\\PlansController::addNewPlan",
                "parameters": [
                    {
                        "name": "broadcasterId",
                        "in": "path",
                        "description": "Id of broadcaster",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "e4c7180e-602d-43d1-b7a9-a0a77c0d49a2"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error description"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "security": [
        []
    ]
}